requirements.txt
pandas~=1.5.3
sqlalchemy~=1.4.39
python-dotenv~=1.2.1
pyarrow